Add files via upload
This is a Quiz application that will have some questions along with four options for each question.
Users can select an option out of the given four options. In case the user selects the correct option a score will be awarded to the user else not.
After all the questions have been answered, the score of the user along with the total percentage of correctly answered questions out of all questions display.